Update README.md
Just spent several hours looking into a bug that really came down to me not presorting the array passed to quantile. Wasn't until the 5th time I went back to read the documentation that I caught that it was a sorted array, so I thought I'd extend the emphasis on array to sorted array.